requirements.txt
streamlit==1.32.0
langchain>=0.1.0
langchain-openai==0.0.5
langchain-community>=0.0.27
openai==1.12.0
tiktoken>=0.5.2,<0.6.0
beautifulsoup4==4.12.3
chromadb
reportlab
python-dotenv==1.0.1
requests==2.31.0
python-dateutil==2.8.2
pytz==2024.1
pandas==2.0.0
plotly==5.18.0
numpy==1.26.4
tiktoken>=0.5.2,<0.6.0